Revert "refact: Move global types to onboarding prototype"
This reverts commit b4c9d7c722f5a3093e1624f68148076ef41d5c96.
