+++ client/globals.d.ts
@@ -4,7 +4,25 @@ declare interface RequirementError {
 	requirement: string;
 }
 
-interface MccsDisplayTreeItem {
+declare interface Country {
+	key: string;
+	name: string;
+	types: BusinessType[];
+}
+
+declare interface BusinessType {
+	key: string;
+	name: string;
+	description: string;
+	structures: BusinessStructure[];
+}
+
+declare interface BusinessStructure {
+	key: string;
+	name: string;
+}
+
+declare interface MccsDisplayTreeItem {
 	id: string;
 	type: string;
 	title: string;
@@ -80,14 +98,7 @@ declare const wcpaySettings: {
 	onboardingTestMode: boolean;
 	onboardingFieldsData?: {
 		business_types: Country[];
-		mccs_display_tree: {
-			id: string;
-			type: string;
-			title: string;
-			items?: MccsDisplayTreeItem[];
-			mcc?: number;
-			keywords?: string[];
-		}[];
+		mccs_display_tree: MccsDisplayTreeItem[];
 	};
 };
