+++ client/onboarding-prototype/form.tsx
@@ -19,6 +19,7 @@ import {
 	PhoneNumberFieldProps,
 } from 'components/form/fields';
 import { useOnboardingContext } from './context';
+import { OnboardingFields } from './types';
 import { useValidation } from './validation';
 import strings from './strings';
 import GroupedSelectControl, {
